+++ crates/miden-lib/src/account/interface/component.rs
@@ -167,7 +167,7 @@ impl AccountComponentInterface {
     ///
     /// ```masm
     ///     push.{note_information}
-    ///     call.::miden::contracts::wallets::basic::create_note
+    ///     call.::miden::tx::create_note
     ///
     ///     push.{note asset}
     ///     call.::miden::contracts::wallets::basic::move_asset_to_note dropw
@@ -241,7 +241,7 @@ impl AccountComponentInterface {
                     // stack => []
                 },
                 AccountComponentInterface::BasicWallet => {
-                    body.push_str("call.::miden::contracts::wallets::basic::create_note\n");
+                    body.push_str("call.::miden::tx::create_note\n");
                     // stack => [note_idx]
 
                     for asset in partial_note.assets().iter() {