+++ crates/miden-lib/src/account/interface/test.rs
@@ -233,24 +233,23 @@ fn test_basic_wallet_custom_notes() {
     let vault = NoteAssets::new(vec![FungibleAsset::mock(100)]).unwrap();
 
     let compatible_source_code = "
+        use.miden::tx
         use.miden::contracts::wallets::basic->wallet
         use.miden::contracts::faucets::basic_fungible->fungible_faucet
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
-                
+
                 # unsupported procs
                 call.fungible_faucet::distribute
                 call.fungible_faucet::burn
             else
                 # supported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
             end
         end
@@ -271,7 +270,7 @@ fn test_basic_wallet_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # unsupported procs
                 call.fungible_faucet::distribute
                 call.fungible_faucet::burn
@@ -280,7 +279,6 @@ fn test_basic_wallet_custom_notes() {
                 call.fungible_faucet::distribute
 
                 # supported procs
-                call.wallet::create_note
                 call.wallet::receive_asset
                 call.wallet::move_asset_to_note
             end
@@ -336,7 +334,7 @@ fn test_basic_fungible_faucet_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported procs
                 call.fungible_faucet::distribute
                 call.fungible_faucet::burn
@@ -346,7 +344,6 @@ fn test_basic_fungible_faucet_custom_notes() {
 
                 # unsupported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
             end
         end
@@ -367,11 +364,11 @@ fn test_basic_fungible_faucet_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported procs
                 call.fungible_faucet::distribute
                 call.fungible_faucet::burn
-            
+
                 # unsupported proc
                 call.wallet::receive_asset
             else
@@ -379,7 +376,6 @@ fn test_basic_fungible_faucet_custom_notes() {
                 call.fungible_faucet::burn
 
                 # unsupported procs
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
             end
         end
@@ -456,7 +452,7 @@ fn test_custom_account_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported proc
                 call.test_account::procedure_1
 
@@ -489,12 +485,11 @@ fn test_custom_account_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 call.wallet::receive_asset
                 call.test_account::procedure_1
             else
                 call.test_account::procedure_2
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
             end
         end
@@ -578,17 +573,15 @@ fn test_custom_account_multiple_components_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
                 call.test_account::procedure_1
                 call.test_account::procedure_2
             else
                 # supported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
                 call.test_account::procedure_1
                 call.test_account::procedure_2
@@ -620,10 +613,9 @@ fn test_custom_account_multiple_components_custom_notes() {
 
         begin
             push.1
-            if.true 
+            if.true
                 # supported procs
                 call.wallet::receive_asset
-                call.wallet::create_note
                 call.wallet::move_asset_to_note
                 call.test_account::procedure_1
                 call.test_account::procedure_2