+++ crates/miden-objects/src/testing/account_component.rs
@@ -12,7 +12,6 @@ use crate::{
 
 pub const BASIC_WALLET_CODE: &str = "
     export.::miden::contracts::wallets::basic::receive_asset
-    export.::miden::contracts::wallets::basic::create_note
     export.::miden::contracts::wallets::basic::move_asset_to_note
 ";