+++ crates/miden-testing/src/kernel_tests/mod.rs
@@ -217,7 +217,7 @@ fn executed_transaction_account_delta_new() -> anyhow::Result<()> {
             # => [tag, aux, note_type, execution_hint, RECIPIENT, pad(8)]
 
             # create the note
-            call.::miden::contracts::wallets::basic::create_note
+            call.tx::create_note
             # => [note_idx, pad(15)]
 
             # move an asset to the created note to partially deplete fungible asset balance
@@ -239,6 +239,7 @@ fn executed_transaction_account_delta_new() -> anyhow::Result<()> {
     let tx_script_src = format!(
         "\
         use.test::account
+        use.miden::tx
 
         ## TRANSACTION SCRIPT
         ## ========================================================================================
@@ -428,6 +429,7 @@ fn test_send_note_proc() -> miette::Result<()> {
         let tx_script_src = format!(
             "\
             use.miden::contracts::wallets::basic->wallet
+            use.miden::tx
             use.test::account
 
             ## TRANSACTION SCRIPT
@@ -445,7 +447,7 @@ fn test_send_note_proc() -> miette::Result<()> {
                 padw padw swapdw
                 # => [tag, aux, execution_hint, note_type, RECIPIENT, pad(8) ...]
 
-                call.wallet::create_note
+                call.tx::create_note
                 # => [note_idx, GARBAGE(15)]
 
                 movdn.4
@@ -586,6 +588,7 @@ fn executed_transaction_output_notes() -> anyhow::Result<()> {
     let tx_script_src = format!(
         "\
         use.miden::contracts::wallets::basic->wallet
+        use.miden::tx
         use.test::account
 
         # Inputs:  [tag, aux, note_type, execution_hint, RECIPIENT]
@@ -596,7 +599,7 @@ fn executed_transaction_output_notes() -> anyhow::Result<()> {
             padw padw swapdw
             # => [tag, aux, execution_hint, note_type, RECIPIENT, pad(8)]
 
-            call.wallet::create_note
+            call.tx::create_note
             # => [note_idx, pad(15)]
 
             # remove excess PADs from the stack